% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostan-package.R
\docType{package}
\name{geostan-package}
\alias{geostan-package}
\alias{geostan}
\title{The 'geostan' package.}
\description{
User-friendly Bayesian geospatial analysis powered by Stan.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
