% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information-criteria.R
\name{log_lik}
\alias{log_lik}
\alias{log_lik.geostan_fit}
\title{Extract log-likelihood}
\usage{
log_lik(object, array = FALSE, ...)

\method{log_lik}{geostan_fit}(object, array = FALSE, ...)
}
\arguments{
\item{object}{A \code{geostan_fit} model}

\item{array}{Return results as an array, one matrix per MCMC chain?}

\item{...}{Other arguments (not used)}
}
\value{
A matrix (or array) of MCMC samples for the log-likelihood: the casewise probability of the data conditional on estimated parameter values.
}
\description{
Extract log-likelihood
}
\seealso{
\code{\link[geostan]{waic}} \code{\link[geostan]{dic}}
}
