% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{mc}
\alias{mc}
\title{The Moran coefficient}
\usage{
mc(x, w)
}
\arguments{
\item{x}{Numeric vector of input values, length n.}

\item{w}{An n x n patial connectivity matrix. See \link[geostan]{shape2mat}.}
}
\value{
The Moran coefficient, a numeric value.
}
\description{
The Moran coefficient, a measure of spatial autocorrelation (also known as Global Moran's I)
}
\examples{
library(sf)
data(turmoil)
w <- shape2mat(turmoil, snap = 1) # snap=1 catches some misaligned borders (WI-IL)
x <- turmoil$unemployment
mc(x, w)

}
