% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{moran_plot}
\alias{moran_plot}
\title{Moran plot}
\usage{
moran_plot(x, w, main = "Moran Plot", cex = 1, lwd = 1,
  ylab = "Spatial Lag", ..., pch = 20, col = "darkred")
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A spatial connectivity matrix.}

\item{main}{Title for the plot.}

\item{cex}{Number indicating the amount by which plotting text and symbols should be scaled relative to the default. Default is 1, 1.5 is 50 percent larger, 0.5 is 50 percent smaller, etc.}

\item{lwd}{linewidth. Default value is 1.}

\item{ylab}{Label for the y-axis. Default is "Spatial Lag".}

\item{...}{additional arugents will be pased to plot.}

\item{pch}{Base R plot symbol for points. Defaults to \code{20}.}

\item{col}{Symbol color. Defaults to \code{darkred}.}
}
\value{
Returns a scatter plot with the given x values on the x-axis and their spatially lagged values on the y-axis (the Moran plot).
}
\description{
Plots a set of values against their mean spatially lagged values and gives the Moran coefficient as a measure of spatial autocorrelation.
}
\examples{
data(turmoil)
x <- turmoil$unemployment
w <- shape2mat(turmoil, snap=1) #snap=1 catches bad borders (IL-WI)
moran_plot(x, w)

}
