% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{n_nbs}
\alias{n_nbs}
\title{Count neighbors in a connectivity matrix}
\usage{
n_nbs(C)
}
\arguments{
\item{C}{A connectivity matrix}
}
\value{
A vector with the number of non-zero values in each row of \code{C}
}
\description{
Count neighbors in a connectivity matrix
}
\examples{

data(sentencing)
C <- shape2mat(sentencing)
sentencing$Ni <- n_nbs(C)

}
