% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostan_fit-methods.R
\name{spatial}
\alias{spatial}
\title{Extract spatial component from a fitted geostan model}
\usage{
spatial(object, summary = TRUE, ...)
}
\arguments{
\item{object}{Fitted geostan model}

\item{summary}{should the posterior distribution be summarized? If \code{FALSE}, returns a matrix of samples; else a \code{data.frame} with summary statistics of the spatial filter at each observation.}

\item{...}{additional arguments}
}
\description{
Extracts the posterior distribution of the spatial component from a fitted geostan model
}
\seealso{
\code{\link[geostan]{geostan_fit}}
}
