% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{student_t}
\alias{student_t}
\title{Student t family}
\usage{
student_t()
}
\value{
An object of class \code{family}
}
\description{
create a family object for the Student t likelihood
}
\examples{
\dontrun{
data(georgia)
fit = stan_glm(log(rate.male) ~ 1, data = georgia, family = student_t())
}
}
