% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turmoil}
\alias{turmoil}
\title{Growth in the Republican Presidential vote share, 2012 to 2016.}
\format{A simple feature collection including the following attributes:
\describe{
 \item{GEOID}{Six digit combined state and county FIPS code}
 \item{state_po}{Two digit state abbreviation}
 \item{county}{County name}
 \item{gop_growth}{Change in the Republican vote share from historic (2000 - 2012) average vote share to 2016 vote share (i.e. (trump_2016/total_2016) - historic_gop)}
 \item{historic_gop}{Average Republican vote share, 2000 to 2012}
 \item{trump_2016}{Number of votes for Donald Trump in 2016}
 \item{total_2016}{Total number of votes in 2016}
 \item{population}{ACS 2016 5 year population estimate (in thousands).}
 \item{college_educated}{ACS 2016 5 year estimate of the percent of the population 25 and older with a bachelors degree or higher}
 \item{white_nonhispanic}{ACS 2016 5 year estimate of the percent of non-hispanic whites in the population}
 \item{unemployment}{Local area unemployment estimate from the Bureau of Labor Statistics, 2016}
}}
\source{
MIT Election Data and Science Lab, 2018, "County Presidential Election Returns 2000-2016", \url{https://doi.org/10.7910/DVN/VOQCHQ}, Harvard Dataverse, V1, UNF:6:ZaxsDvp6RsFitno8ZYlK7w== [fileUNF]

US Bureau of Labor Statistics, Local Force By County, 2016 Annual Averages. \url{https://www.bls.gov/lau/#cntyaa}

US Census Bureau, 2016. American Community Survey. B01003, B01003, B15003.

Kyle Walker, 2018. Tidycensus. \url{https://walkerke.github.io/tidycensus/index.html}
}
\usage{
turmoil
}
\description{
A simple features spatial dataset containing county attributes and 2016 Presidential election data for Illinois, Indiana, Michigan, Ohio, Pennsylvania and Wisconsin.
 This is a processed version of the MIT "County Presidential Election Returns 2000-2016" data set augmented with data from the Census Bureau and Bureau of Labor Statistics.
}
\examples{
\dontrun{
library(sf)
data(turmoil)
}
}
\keyword{datasets}
