% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostan_fit-methods.R
\name{print.geostan_fit}
\alias{print.geostan_fit}
\alias{plot.geostan_fit}
\alias{as.matrix.geostan_fit}
\alias{as.data.frame.geostan_fit}
\alias{as.array.geostan_fit}
\alias{residuals.geostan_fit}
\alias{fitted.geostan_fit}
\alias{spatial.geostan_fit}
\alias{posterior_predict.geostan_fit}
\title{geostan_fit methods}
\usage{
\method{print}{geostan_fit}(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), digits = 3, ...)

\method{plot}{geostan_fit}(x, pars, plotfun = "stan_plot", ...)

\method{as.matrix}{geostan_fit}(x, ...)

\method{as.data.frame}{geostan_fit}(x, ...)

\method{as.array}{geostan_fit}(x, ...)

\method{residuals}{geostan_fit}(object, summary = TRUE, ...)

\method{fitted}{geostan_fit}(object, summary = TRUE, ...)

\method{spatial}{geostan_fit}(object, summary = TRUE, ...)

\method{posterior_predict}{geostan_fit}(object, draws, summary = FALSE, ...)
}
\arguments{
\item{x}{A fitted model object of class \code{geostan_fit}.}

\item{probs}{Argument passed to \code{quantile}; which quantiles to calculate and print.}

\item{digits}{number of digits to print}

\item{...}{additional arguments.}

\item{pars}{parameters to include; a character string or vector of parameter names.}

\item{plotfun}{Argument passed to \code{rstan::plot}. Options include "hist", "trace", and "dens".}

\item{object}{A fitted model object of class \code{geostan_fit}.}

\item{summary}{Logical (default \code{summary = TRUE}); should the values be summarized with the mean, standard deviation and quantiles (\code{probs = c(.025, .2, .5, .8, .975)}) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.}

\item{draws}{Number of samples to return; maximum and default value is the number of MCMC samples stored by the model.}
}
\value{
return either a matrix containing all samples for each observation, or a \code{data.frame} containing a summary values for each observation.
}
\description{
Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.
}
\examples{
\dontrun{
library(geostan)
library(sf)
data(turmoil)
ohio <- turmoil[which(state_po == "OH"), ]
fit <- stan_esf(gop_growth ~ historic_gop + log(population), shape = ohio, chains = 1, iter = 600)

# print and plot results
print(fit)
plot(fit)
plot(fit, pars = c("alpha", "beta")
plot(fit, pars = c("b_historic_gop", "b_log(population)"), plotfun = "hist", bins = 30)
plot(fit, plotfun = "trace")

# extract posterior samples
as.data.frame(fit)
as.data.frame(fit, pars = "beta")
as.data.frame(fit, pars = "b_historic_gop")
as.array(fit)
as.matrix(fit)

# extract residuals, etc.
res <- residuals(fit)
##res <- resid(fit) # alias
head(res)
plot(ohio$gop_growth, res$mean)
mat <- resid(fit, summary = FALSE)
dim(mat)
pp <- posterior_predict(fit, summary  = FALSE)
dim(pp)
head(fitted(fit))

# extract and plot the mean estimate of the spatial filter
ohio$sf <- spatial(fit, summary = TRUE)$mean
plot(ohio[,'sf'])
}
}
\seealso{
\link[geostan]{stan_esf} \link[rstan]{stan_plot}
}
