% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostan_fit-methods.R
\name{geostan_fit}
\alias{geostan_fit}
\alias{print.geostan_fit}
\alias{plot.geostan_fit}
\alias{as.matrix.geostan_fit}
\alias{as.data.frame.geostan_fit}
\alias{as.array.geostan_fit}
\alias{residuals.geostan_fit}
\alias{fitted.geostan_fit}
\alias{spatial.geostan_fit}
\title{geostan_fit methods}
\usage{
\method{print}{geostan_fit}(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), digits = 3, ...)

\method{plot}{geostan_fit}(x, pars, plotfun = "dens", ...)

\method{as.matrix}{geostan_fit}(x, ...)

\method{as.data.frame}{geostan_fit}(x, ...)

\method{as.array}{geostan_fit}(x, ...)

\method{residuals}{geostan_fit}(object, summary = TRUE, ...)

\method{fitted}{geostan_fit}(object, summary = TRUE, ...)

\method{spatial}{geostan_fit}(object, summary = TRUE, ...)
}
\arguments{
\item{x}{A fitted model object of class \code{geostan_fit}.}

\item{probs}{Argument passed to \code{quantile}; which quantiles to calculate and print.}

\item{digits}{number of digits to print}

\item{...}{additional arguments.}

\item{pars}{parameters to include; a character string or vector of parameter names.}

\item{plotfun}{Argument passed to \code{rstan::plot}. Options include "hist", "trace", and "dens".}

\item{object}{A fitted model object of class \code{geostan_fit}.}

\item{summary}{Logical (default \code{summary = TRUE}); should the values be summarized with the mean, standard deviation and quantiles (\code{probs = c(.025, .2, .5, .8, .975)}) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.}
}
\value{
return either a matrix containing all samples for each observation, or a \code{data.frame} containing a summary values for each observation.
}
\description{
Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)
data(ohio)
fit <- stan_esf(gop_growth ~ historic_gop + log(pop_density),
               data = ohio,
               scalex = TRUE,
               C = shape2mat(ohio),
               chains = 3,
               iter = 1e3,
               refresh = 0 # less printing
)

# print and plot results
print(fit)
plot(fit)

beta.samples <- as.matrix(fit, pars = "beta")
beta.mean <- apply(beta.samples, 2, mean)
print(beta.mean)

# extract residuals and fitted values
res <- residuals(fit)
res <- resid(fit) 
head(res)
f <- fitted(fit)
plot(f$mean, res$mean, xlab = "fitted values", ylab = "residuals")
mat <- resid(fit, summary = FALSE)
dim(mat)

# extract and plot the posterior mean of the spatial filter
sf <- spatial(fit, summary = TRUE)
head(sf)
ohio$sf <- sf$mean
ggplot(ohio) +
  geom_sf(aes(fill = sf)) +
  scale_fill_gradient2()  
}

}
\seealso{
\link[geostan]{stan_esf} \link[rstan]{stan_plot}
}
