% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostan_fit-methods.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Draw samples from posterior predictive distribution of a fitted geostan model}
\usage{
posterior_predict(object, draws, summary = FALSE, ...)
}
\arguments{
\item{object}{A fitted geostan model}

\item{draws}{Number of samples to return; maximum and default value is the number of MCMC samples stored by the model.}

\item{summary}{Logical (default \code{summary = TRUE}); should the values be summarized with the mean, standard deviation and quantiles (\code{probs = c(.025, .2, .5, .8, .975)}) for each observation? Otherwise a matrix containing samples for each observation is returned.}

\item{...}{additional arguments.}
}
\description{
Draws samples from the posterior predictive distribution of a fitted geostan model
}
