% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Draw samples from the posterior predictive distribution}
\usage{
posterior_predict(
  object,
  newdata,
  C,
  samples,
  predictive = TRUE,
  re_form = NULL,
  spatial = TRUE,
  summary = FALSE,
  width = 0.95,
  seed,
  centerx = FALSE,
  scalex = FALSE
)
}
\arguments{
\item{object}{A \code{geostan_fit} object.}

\item{newdata}{A \code{data.frame} or \code{matrix} with data to use for predictions. Must be named and the names must match the model formula. If the model has an slx term and \code{spatial = TRUE} then \code{newdata} must have the same number of rows as the data used to fit the model.}

\item{C}{Spatial connectivity matrix for calculating spatially lagged covariates (slx). Only used if \code{spatial = TRUE} and the model contains an \code{slx} term. It will be row-standardized automatically.}

\item{samples}{Number of samples to take from the posterior distribution. Maximum is the total number of samples stored in the model.}

\item{predictive}{Return the predictive distribution? Defaults to \code{TRUE}. If \code{FALSE}, then the linear predictor is returned after applying the inverse link function (results are on the same scale as the outcome variable).}

\item{re_form}{If \code{re_form = NA} any random effects terms will be ignored when making predictions. If default \code{re_form = NULL} is set, then the results will include random effects terms. Those are the only options; this syntax is used to maintain some consistency with other R packages which have additional options.}

\item{spatial}{Include the spatial component in the model predictions? Defaults to \code{TRUE} and will be ignored if the model does not have a spatial comonent (i.e. \code{\link[geostan]{stan_glm}}). For models fit by \code{\link[geostan]{stan_esf}}, \code{\link[geostan]{stan_icar}}, and \code{\link[geostan]{stan_car}} this option requires that \code{newdata} have the same number of observations as the data that the model was fit to.}

\item{summary}{Should the posterior predictions be summarized using quantile intervals? It \code{summary = TRUE} the function returns a data frame with the expected value of the posterior distribution (column \code{mu}) and 100*width credible intervals for, respectively, the posterior predictive distribution (columns \code{pred.lwr}, \code{pred.upr}) and mean outcome (columns \code{mu.lwr}, \code{mu.upr}).}

\item{width}{Only used if \code{summary = TRUE} to set the quantiles for the credible intervals.}

\item{seed}{A single integer value to be used in a call to \code{\link[base]{set.seed}} before taking samples from the posterior distribution. Passing a value to \code{seed} enables you to obtain the same results each time (by using the same seed).}

\item{centerx}{Should \code{newdata} be centered using the means of the variables in the original data used to fit the model (stored in \code{object$scale_params})? Defaults to \code{FALSE}.}

\item{scalex}{Should \code{newdata} be centered and scaled using the standard deviations of the variables in the original data used to fit the model (stored in \code{object$scale_params})? Defaults to \code{FALSE}.}
}
\value{
A matrix of size \code{S} x \code{N} containing samples from the posterior predictive distribution, where \code{S} is the number of samples and \code{N} is the number of observations. It is of class \code{matrix} and \code{ppd}.

If \code{summary = TRUE} the return value is a data frame with the expected value of the posterior distribution (column \code{mu}) and 100*width credible intervals for, respectively, the posterior predictive distribution (columns \code{pred.lwr}, \code{pred.upr}) and mean outcome (columns \code{mu.lwr}, \code{mu.upr}). The width of the intervals and the quantiles corresponding to the upper and lower limits of the credible intervals are returned as an attribute of the data frame (use \code{attributes(df)$interval} to retrieve them, where \code{df} is the data frame returned by \code{posterior_predict}).
}
\description{
Draw samples from the posterior predictive distribution of a fitted \code{geostan} model. Use the original data or new data, with or without the spatial component and random effects terms.
}
\examples{
library(bayesplot)
library(ggplot2)
library(sf)
data(ohio)

## fit Ohio election model with a spatial filter
fit <- stan_esf(gop_growth ~ 1,
                data = ohio,
                C = shape2mat(ohio),
                chains = 1,  # chains = 4
                iter = 200, # iter = 2e3
                cores = 1) # cores = 4

# posterior predictive check (density overlay)
 # compare distribution of observations to the predictive distribution
y <- ohio$gop_growth
yrep <- posterior_predict(fit, samples = 75)
ppc_dens_overlay(y, yrep)

## plot marginal effects with credible intervals 
fit <- stan_esf(gop_growth ~ poly(college_educated, df = 3),
                data = ohio,
                C = shape2mat(ohio),
                cores = 1,  # cores = 4,
                chains = 1, # chains = 4
                iter = 200) # iter = 2e3

x <- seq(10, 45, by = 1)
newdata <- data.frame(college_educated = x)
pred.df <- posterior_predict(fit, newdata, spatial = FALSE, summary = TRUE)

## plot the mean of the predictive distribution (regression line)
a <- ggplot(pred.df, aes(x = mu)) +
    geom_line(aes( y = mu))

## add the credible interval for the mean
a <- a +
    geom_ribbon(aes(ymin = mu.lwr,
                    ymax = mu.upr),
               fill = "blue", apha = 0.5)

## add the credible interval for predicted values
a +
  geom_ribbon(aes(ymin = pred.lwr,
                  ymax = pred.upr),
              fill = "blue", alpha = 0.5)


}
