% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_icar.R
\name{stan_icar}
\alias{stan_icar}
\title{Intrinsic autoregressive models}
\source{
Besag, J. (1974). Spatial interaction and the statistical analysis of lattice systems. Journal of the Royal Statistical Society: Series B (Methodological), 36(2), 192-225.

Besag, J., York, J., & Mollié, A. (1991). Bayesian image restoration, with two applications in spatial statistics. Annals of the institute of statistical mathematics, 43(1), 1-20.

Donegan, Connor. 2021. Flexible functions for ICAR, BYM, and BYM2 models in Stan. Code repository. \url{https://github.com/ConnorDonegan/Stan-IAR}

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure. \emph{Int. J. Env. Res. and Public Health} 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: \url{https://github.com/ConnorDonegan/survey-HBM}.

Donegan, Connor (2021). Spatial conditional autoregressive models in Stan. \emph{OSF Preprints}. \doi{10.31219/osf.io/3ey65}.

Freni-Sterrantino, Anna, Massimo Ventrucci, and Håvard Rue. 2018. A Note on Intrinsic Conditional Autoregressive Models for Disconnected Graphs. Spatial and Spatio-Temporal Epidemiology 26: 25–34.

Morris, M., Wheeler-Martin, K., Simpson, D., Mooney, S. J., Gelman, A., & DiMaggio, C. (2019). Bayesian hierarchical spatial models: Implementing the Besag York Mollié model in stan. Spatial and spatio-temporal epidemiology, 31, 100301.

Riebler, A., Sorbye, S. H., Simpson, D., & Rue, H. (2016). An intuitive Bayesian spatial model for disease mapping that accounts for scaling. Statistical Methods in Medical Research, 25(4), 1145-1165.
}
\usage{
stan_icar(
  formula,
  slx,
  re,
  data,
  C,
  family = poisson(),
  type = c("icar", "bym", "bym2"),
  scale_factor = NULL,
  prior = NULL,
  ME = NULL,
  centerx = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 500,
  pars = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{formula}{A model formula, following the R \link[stats]{formula} syntax. Binomial models can be specified by setting the left hand side of the equation to a data frame of successes and failures, as in \code{cbind(successes, failures) ~ x}.}

\item{slx}{Formula to specify any spatially-lagged covariates. As in, \code{~ x1 + x2} (the intercept term will be removed internally). When setting priors for \code{beta}, remember to include priors for any SLX terms.}

\item{re}{To include a varying intercept (or "random effects") term, \code{alpha_re}, specify the grouping variable here using formula syntax, as in \code{~ ID}. Then, \code{alpha_re} is a vector of parameters added to the linear predictor of the model, and:\preformatted{       alpha_re ~ N(0, alpha_tau)
       alpha_tau ~ Student_t(d.f., location, scale).
}

Before using this term, read the \code{Details} section and the \code{type} argument. Specifically, if you use \code{type = bym}, then an observational-level \code{re} term is already included in the model. (Similar for \code{type = bym2}.)}

\item{data}{A \code{data.frame} or an object coercible to a data frame by \code{as.data.frame} containing the model data.}

\item{C}{Spatial connectivity matrix which will be used to construct an edge list for the ICAR model, and to calculate residual spatial autocorrelation as well as any user specified \code{slx} terms. It will automatically be row-standardized before calculating \code{slx} terms. \code{C} must be a binary symmetric \code{n x n} matrix.}

\item{family}{The likelihood function for the outcome variable. Current options are \code{binomial(link = "logit")} and \code{poisson(link = "log")}.}

\item{type}{Defaults to "icar" (partial pooling of neighboring observations through parameter \code{phi}); specify "bym" to add a second parameter vector \code{theta} to perform partial pooling across all observations; specify "bym2" for the innovation introduced by Riebler et al. (2016). See \code{Details} for more information.}

\item{scale_factor}{For the BYM2 model, optional. If missing, this will be set to a vector of ones. See \code{Details}.}

\item{prior}{A named list of parameters for prior distributions (see \code{\link[geostan]{priors}}):
\describe{

\item{intercept}{The intercept is assigned a Gaussian prior distribution (see \code{\link[geostan]{normal}}}.

\item{beta}{Regression coefficients are assigned Gaussian prior distributions. Variables must follow their order of appearance in the model \code{formula}. Note that if you also use \code{slx} terms (spatially lagged covariates), and you use custom priors for \code{beta}, then you have to provide priors for the slx terms. Since slx terms are \emph{prepended} to the design matrix, the prior for the slx term will be listed first.
}

\item{sigma}{For \code{family = gaussian()} and \code{family = student_t()} models, the scale parameter, \code{sigma}, is assigned a (half-) Student's t prior distribution. The half-Student's t prior for \code{sigma} is constrained to be positive.}

\item{nu}{\code{nu} is the degrees of freedom parameter in the Student's t likelihood (only used when \code{family = student_t()}). \code{nu} is assigned a gamma prior distribution. The default prior is \code{prior = list(nu = gamma(alpha = 3, beta = 0.2))}.
}

\item{tau}{The scale parameter for random effects, or varying intercepts, terms. This scale parameter, \code{tau}, is assigned a half-Student's t prior. To set this, use, e.g., \code{prior = list(tau = student_t(df = 20, location = 0, scale = 20))}.}
}}

\item{ME}{To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list of priors. See the Details section below for more information. Elements of the \code{ME} list may include:
\describe{

\item{se}{A required dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of \code{model.matrix(formula, data)}.}

\item{bounds}{An optional numeric vector of length two providing the upper and lower bounds, respectively, of the variables. If not provided, they will be set to \code{c(-Inf, Inf)} (i.e., unbounded). Common usages include keeping percentages between zero and one hundred or proportions between zero and one.}

\item{car_parts}{A list of data for the CAR model, as returned by \code{\link[geostan]{prep_car_data}}. If not provided, a non-spatial Student's t model will be used instead of the CAR model.}

\item{prior}{Optionally provide a named list of prior distributions for the measurement error model(s) (see \code{\link[geostan]{priors}}). If none are provided, default priors will be assigned and printed to the console. \describe{

\item{df}{If using a non-spatial ME model, the degrees of freedom for the Student's t model is assigned a gamma prior with default parameters of \code{gamma(alpha = 3, beta = 0.2)}. Provide values for each covariate in \code{se}, listing the values in the same order as the columns of \code{se}.}

\item{car_rho}{The CAR model also has a spatial autocorrelation parameter, \code{rho}, which is assigned a uniform prior distribution. You can set the boundaries of the prior with: \code{ME$prior$car_rho <- uniform(lower_bound, upper_bound)}. You must specify values that are within the permissible range of values for \code{rho}. This range is automatically printed to the console by \code{\link[geostan]{prep_car_data}}.}

\item{location}{The prior for the location parameter, mu, is Gaussian (the default being \code{normal(location = 0, scale = 100)}). To adjust the prior distributions, provide values for each covariate in \code{se}, listing the values in the same order as the columns of \code{se}.}

\item{scale}{The prior for the \code{scale} parameters is Student's t, and the default is \code{student_t(df = 10, location = 0, scale = 40)}. To adjust, provide values for each covariate in \code{se}, listing the values in the same order as the columns of \code{se}.}
}
}
}}

\item{centerx}{To center predictors on their mean values, use \code{centerx = TRUE}. If \code{centerx} is a numeric vector, predictors will be centered on the values provided. This argument is passed to \code{\link[base]{scale}}.}

\item{prior_only}{Draw samples from the prior distributions of parameters only.}

\item{chains}{Number of MCMC chains to estimate.}

\item{iter}{Number of samples per chain. .}

\item{refresh}{Stan will print the progress of the sampler every \code{refresh} number of samples; set \code{refresh=0} to silence this.}

\item{pars}{Optional; specify any additional parameters you'd like stored from the Stan model.}

\item{control}{A named list of parameters to control the sampler's behavior. See \code{\link[rstan]{stan}} for details.}

\item{...}{Other arguments passed to \link[rstan]{sampling}. For multi-core processing, you can use \code{cores = parallel::detectCores()}, or run \code{options(mc.cores = parallel::detectCores())} first.}
}
\value{
An object of class class \code{geostan_fit} (a list) containing:
\describe{
\item{summary}{Summaries of the main parameters of interest; a data frame}
\item{diagnostic}{Widely Applicable Information Criteria (WAIC) with a measure of effective number of parameters (\code{eff_pars}) and mean log pointwise predictive density (\code{lpd}), and mean residual spatial autocorrelation as measured by the Moran coefficient.}
\item{stanfit}{an object of class \code{stanfit} returned by \code{rstan::stan}}
\item{data}{a data frame containing the model data}
\item{edges}{The edge list representing all unique sets of neighbors and the weight attached to each pair (i.e., their corresponding element in the connectivity matrix  C}
\item{family}{the user-provided or default \code{family} argument used to fit the model}
\item{formula}{The model formula provided by the user (not including ICAR component)}
\item{slx}{The \code{slx} formula}
\item{re}{A list with two name elements, \code{formula} and \code{Data}, containing the formula \code{re} and a data frame with columns \code{id} (the grouping variable) and \code{idx} (the index values assigned to each group).}
\item{priors}{Prior specifications.}
\item{x_center}{If covariates are centered internally (i.e., \code{centerx} is not \code{FALSE}), then \code{x_centers} is the numeric vector of values on which the covariates were centered.}
\item{spatial}{A data frame with the name of the spatial parameter (\code{"phi"} if \code{type = "icar"} else \code{"convolution"}) and method (\code{toupper(type)}).}
}
}
\description{
The intrinsic conditional auto-regressive (ICAR) model for spatial count data. Options include the BYM model, the BYM2 model, and a solo ICAR term.
}
\details{
The Stan code for the ICAR component of the model and the BYM2 option is from Morris et al. (2019) with adjustments to enable non-binary weights and disconnected graph structures (see Freni-Sterrantino (2018) and Donegan (2021)).

The exact specification depends on the \code{type} argument.
\subsection{'icar'}{

For Poisson models for count data, y, the basic model specification (\code{type = "icar"}) is:\preformatted{        y ~ Poisson(exp(offset + mu + phi))
        phi ~ ICAR(spatial_scale)
        spatial_scale ~ Gaussian(0, 1)
}

where \code{mu} contains an intercept and potentially covariates. The spatial trend, \code{phi}, has a mean of zero and a single scale parameter, \code{spatial_scale}.

The ICAR prior model is a CAR model that has a spatial autocorrelation parameter \code{car_alpha} equal to 1 (see \link[geostan]{stan_car}). Thus the ICAR prior places high probability on a smooth spatially (or temporally) varying mean. This is rarely sufficient to model the amount of variation present in social and health data.
}

\subsection{'bym'}{

Often, an observational-level random effect term, \code{theta}, is added to capture (heterogeneous or unstructured) deviations from \code{mu + phi}. The combined term is referred to as a convolution term:\preformatted{        convolution = phi + theta.
}

This is known as the BYM model (Besag et al. 1991), and can be specified using \code{type = "bym"}:\preformatted{        y ~ Poisson(exp(offset + mu + phi + theta))
        phi ~ ICAR(spatial_scale)
        theta ~ Gaussian(0, theta_scale)
        spatial_scale ~ Gaussian(0, 1)
        theta_scale ~ Gaussian(0, 1)
}
}

\subsection{'bym2'}{

Riebler et al. (2016) introduce a variation on the BYM model (\code{type = "bym2"}). This specification combines \code{phi} and \code{theta} using a mixing parameter, \code{rho}, that controls the proportion of the variation that is attributable to the spatially autocorrelated term, \code{phi}, rather than the spatially unstructured term, \code{theta}. The terms share a single scale parameter:\preformatted{        convolution = [sqrt(rho/scale_factor) * phi_tilde + sqrt(1 - rho) * theta_tilde] * spatial_scale.
        phi_tilde ~ Gaussian(0, 1)
        theta_tilde ~ Gaussian(0, 1)
        spatial_scale ~ Gaussian(0, 1)
}

The two \verb{_tilde} terms are standard normal deviates, \code{rho} is restricted to values between zero and one, and \code{scale_factor} is a constant term provided by the user. By default, \code{scale_factor} is equal to one, so that it does nothing. Riebler et al. (2016) argue that the interpretation or meaning of the scale of the ICAR model depends on the graph structure, \code{C}. This implies that the same prior distribution assigned to the \code{spatial_scale} will differ in its implications if \code{C} is changed; in other words, the priors are not transportable across models, and models that use the same nominal prior actually have different priors assigned to \code{spatial_scale}.

Borrowing \code{R} code from Morris (2017) and following Freni-Sterrantino et al. (2018), the following \code{R} code can be used to create the \code{scale_factor} for the BYM2 model (note, this requires the INLA R package), given a spatial adjacency matrix, C:\preformatted{## create a list of data for stan_icar
icar.data <- geostan::prep_icar_data(C)
## calculate scale_factor for each of k connected group of nodes
k <- icar.data$k
scale_factor <- vector(mode = "numeric", length = k)
for (j in 1:k) \{
  g.idx <- which(icar.data$comp_id == j) 
  if (length(g.idx) == 1) \{
       scale_factor[j] <- 1
       next
    \}    
  Cg <- C[g.idx, g.idx] 
  scale_factor[j] <- scale_c(Cg) 
\}
}

This code adjusts for 'islands' or areas with zero neighbors, and it also handles disconnected graph structures (see Donegan 2021). Following Freni-Sterrantino (2018), disconnected components of the graph structure are given their own intercept term; however, this value is added to \code{phi} automatically inside the Stan model. Therefore, the user never needs to make any adjustments for this term. (If you want to avoid complications from a disconnected graph structure, see \code{\link[geostan]{stan_car}}).

Note, the code above requires the \code{scale_c} function; it has package dependencies that are not included in \code{geostan}. To use \code{scale_c}, you have to load the following \code{R} function:\preformatted{#' compute scaling factor for adjacency matrix, accounting for differences in spatial connectivity 
#'
#' @param C connectivity matrix
#'
#' @details
#'
#' Requires the following packages: 
#'
#' library(Matrix)
#' library(INLA);
#' library(spdep)
#' library(igraph)
#'  
#' @source
#'
#'   Morris, Mitzi (2017). Spatial Models in Stan: Intrinsic Auto-Regressive Models for Areal Data. <https://mc-stan.org/users/documentation/case-studies/icar_stan.html>
#'
scale_c <- function(C) \{
 geometric_mean <- function(x) exp(mean(log(x))) 
 N = dim(C)[1]
 Q =  Diagonal(N, rowSums(C)) - C
 Q_pert = Q + Diagonal(N) * max(diag(Q)) * sqrt(.Machine$double.eps)
 Q_inv = inla.qinv(Q_pert, constr=list(A = matrix(1,1,N),e=0))
 scaling_factor <- geometric_mean(Matrix::diag(Q_inv)) 
 return(scaling_factor) 
\}
}
}

\subsection{Spatially lagged covariates (SLX)}{

The \code{slx} argument is a convenience function for including SLX terms. For example,\preformatted{stan_glm(y ~ x1 + x2, slx = ~ x1, ...)
}

is a shortcut for\preformatted{stan_glm(y ~ I(W \%*\% x1) + x1 + x2, ...)
}

where \code{W} is a row-standardized spatial weights matrix (see \code{\link[geostan]{shape2mat}}). SLX terms will always be \emph{prepended} to the design matrix, as above, which is important to know when setting prior distributions for regression coefficients.

For measurement error (ME) models, the SLX argument is the only way to include spatially lagged covariates since the SLX term needs to be re-calculated on each iteration of the MCMC algorithm.
}

\subsection{Measurement error (ME) models}{

The ME models are designed for surveys with spatial sampling designs, such as the American Community Survey (ACS) estimates (Donegan et al. 2021; Donegan 2021). With estimates, \code{x}, and their standard errors, \code{se}, the ME models have one of the the following two specifications, depending on the user input:\preformatted{       x ~ Gauss(x_true, se)
       x_true ~ MVGauss(mu, Sigma)
       Sigma = (I - rho * C)^(-1) M * tau^2
       mu ~ Gauss(0, 100)
       tau ~ student_t(10, 0, 40)
       rho ~ uniform(lower_bound, upper_bound)
}

where the covariance matrix, \code{Sigma}, has the conditional autoregressive specification, and \code{tau} is the scale parameter. If \code{ME$car_parts} is not provided by the user, then a non-spatial model will be used instead:\preformatted{       x ~ Gauss(x_true, se)
       x_true ~ student_t(df, mu, sigma)
       df ~ gamma(3, 0.2)
       mu ~ Gauss(0, 100)
       sigma ~ student_t(10, 0, 40)
}

The observational error is the difference between the survey estimate and \code{x_true}, the actual value of the variable during the period of the survey.
}
}
\examples{
\dontrun{
library(rstan)
library(bayesplot)
library(sf)
options(mc.cores = parallel::detectCores())
data(sentencing)

C <- shape2mat(sentencing, "B")
log_e <- log(sentencing$expected_sents)
fit.bym <- stan_icar(sents ~ offset(log_e),
                     family = poisson(),
                     data = sentencing,
                     type = "bym",
                     C = C
 )

# check effective sample size and convergence
rstan::stan_ess(fit.bym$stanfit)
rstan::stan_rhat(fit.bym$stanfit)

# see some spatial diagnostics
sp_diag(fit.bym, sentencing)

# posterior predictive distribution
yrep <- posterior_predict(fit.bym, S = 100)
y <- sentencing$sents
bayesplot::ppc_dens_overlay(y, yrep)

# map the smooth spatial term
sp.trend <- spatial(fit.bym)$mean
ggplot( st_as_sf(sentencing) ) +
  geom_sf(aes(fill = sp.trend)) +
  scale_fill_gradient2() +
  theme_void()

# calculate log-standardized sentencing ratios (log-SSRs)
## (like Standardized Incidence Ratios: observed/exected case counts)
f <- fitted(fit.bym)$mean
SSR <- f / sentencing$expected_sents
log.SSR <- log( SSR, base = 2)

ggplot( st_as_sf(sentencing) ) +
  geom_sf(aes(fill = log.SSR)) +
  scale_fill_gradient2() +
  labs(title = "Log-standardized sentencing ratios",
       subtitle = "log( Fitted/Expected), base 2") +
  theme_void() +
  theme(
   legend.position = "bottom",
   legend.key.height = unit(0.35, "cm"),
   legend.key.width = unit(1.5, "cm")
  )
}

}
\seealso{
\link[geostan]{shape2mat}, \link[geostan]{stan_car}, \link[geostan]{stan_esf}, \link[geostan]{stan_glm}, \link[geostan]{prep_icar_data}
}
\author{
Connor Donegan, \email{Connor.Donegan@UTDallas.edu}
}
